Balance Tweaks
As discussed, taking the nerf bat to some of these recipies. If we want people to actually use spacecraft we need to stop them being so wretchedly hard to build, a series of pipes with fluid in them to make a radiator SHOULD NOT require 5 advanced components for gods sake.